refactor(assembler): fix arg order of compile_and_statically_link
The previous order was inconsistent with
`Assembler::assemble_library_from_dir`, and was a bit confusing, in that
the method name implies that the directory is the most important
argument, but that argument came last.

This commit makes the order of the path/namespace argument consistent in
the assembler and parser.