refactor: do not use full constant expression repr in exports
This commit ensures that the type of a constant value present in
library/package exports represents _only_ a concrete value, and not the
full const expression syntax, which is unnecessarily complex to handle
downstream, and would imply a bunch of corner cases that shouldn't exist
in practice anyway.

We instead define `ConstantValue` to represent the subset of expressions
which correspond to a literal value, and export that instead.