refactor: simplify linker and implement link-time const evaluation in the assembler (#2370)
This addresses an issue identified during code review whereby
inter-module constant references might be incorrectly treated as
containing an unresolvable reference to another constant, even though
the reference is in fact resolvable, just not evaluated to a concrete
value yet.

The cause was assuming that the ConstEnvironment::get method returned a
fully-evaluated ConstantExpr that would only be occupied by a Var or
BinaryExpr variant if the expression contained a symbol reference that
could not be resolved in the current environment. This is not actually
the case - the ConstEnvironment::get method returns whatever
ConstantExpr is bound to a given symbol name or path, and only if the
evaluation of that expression was cached would it be safe to assume that
the expression represents a value in all cases.

This commit updates the ConstEvalVisitor to always attempt evaluating
unfolded/partially-evaluated ConstantExprs returned by
ConstEnviroment::get before treating it as unresolvable in that
environment.